Remove all references to IBootstrapperApplication and IBootstrapperEngine from the engine.
