+++ src/burn/engine/EngineForApplication.cpp
@@ -798,439 +798,6 @@ LExit:
     return hr;
 }
 
-class CEngineForApplication : public IBootstrapperEngine, public IMarshal
-{
-public: // IUnknown
-    virtual STDMETHODIMP QueryInterface(
-        __in const IID& riid,
-        __out void** ppvObject
-        )
-    {
-        HRESULT hr = S_OK;
-
-        ExitOnNull(ppvObject, hr, E_INVALIDARG, "Invalid argument ppvObject");
-        *ppvObject = NULL;
-
-        if (::IsEqualIID(__uuidof(IBootstrapperEngine), riid))
-        {
-            *ppvObject = static_cast<IBootstrapperEngine*>(this);
-        }
-        else if (::IsEqualIID(IID_IMarshal, riid))
-        {
-            *ppvObject = static_cast<IMarshal*>(this);
-        }
-        else if (::IsEqualIID(IID_IUnknown, riid))
-        {
-            *ppvObject = reinterpret_cast<IUnknown*>(this);
-        }
-        else // no interface for requested iid
-        {
-            ExitFunction1(hr = E_NOINTERFACE);
-        }
-
-        AddRef();
-
-    LExit:
-        return hr;
-    }
-
-    virtual STDMETHODIMP_(ULONG) AddRef()
-    {
-        return ::InterlockedIncrement(&this->m_cReferences);
-    }
-
-    virtual STDMETHODIMP_(ULONG) Release()
-    {
-        long l = ::InterlockedDecrement(&this->m_cReferences);
-        if (0 < l)
-        {
-            return l;
-        }
-
-        delete this;
-        return 0;
-    }
-
-public: // IBootstrapperEngine
-    virtual STDMETHODIMP GetPackageCount(
-        __out DWORD* /*pcPackages*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP GetVariableNumeric(
-        __in_z LPCWSTR /*wzVariable*/,
-        __out LONGLONG* /*pllValue*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP GetVariableString(
-        __in_z LPCWSTR /*wzVariable*/,
-        __out_ecount_opt(*pcchValue) LPWSTR /*wzValue*/,
-        __inout DWORD* /*pcchValue*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP GetVariableVersion(
-        __in_z LPCWSTR /*wzVariable*/,
-        __out DWORD64* /*pqwValue*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP FormatString(
-        __in_z LPCWSTR /*wzIn*/,
-        __out_ecount_opt(*pcchOut) LPWSTR /*wzOut*/,
-        __inout DWORD* /*pcchOut*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP EscapeString(
-        __in_z LPCWSTR /*wzIn*/,
-        __out_ecount_opt(*pcchOut) LPWSTR /*wzOut*/,
-        __inout DWORD* /*pcchOut*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP EvaluateCondition(
-        __in_z LPCWSTR /*wzCondition*/,
-        __out BOOL* /*pf*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP Log(
-        __in BOOTSTRAPPER_LOG_LEVEL /*level*/,
-        __in_z LPCWSTR /*wzMessage*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SendEmbeddedError(
-        __in DWORD /*dwErrorCode*/,
-        __in_z_opt LPCWSTR /*wzMessage*/,
-        __in DWORD /*dwUIHint*/,
-        __out int* /*pnResult*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SendEmbeddedProgress(
-        __in DWORD /*dwProgressPercentage*/,
-        __in DWORD /*dwOverallProgressPercentage*/,
-        __out int* /*pnResult*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SetUpdate(
-        __in_z_opt LPCWSTR /*wzLocalSource*/,
-        __in_z_opt LPCWSTR /*wzDownloadSource*/,
-        __in DWORD64 /*qwSize*/,
-        __in BOOTSTRAPPER_UPDATE_HASH_TYPE /*hashType*/,
-        __in_bcount_opt(cbHash) BYTE* /*rgbHash*/,
-        __in DWORD /*cbHash*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SetLocalSource(
-        __in_z LPCWSTR /*wzPackageOrContainerId*/,
-        __in_z_opt LPCWSTR /*wzPayloadId*/,
-        __in_z LPCWSTR /*wzPath*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SetDownloadSource(
-        __in_z LPCWSTR /*wzPackageOrContainerId*/,
-        __in_z_opt LPCWSTR /*wzPayloadId*/,
-        __in_z LPCWSTR /*wzUrl*/,
-        __in_z_opt LPCWSTR /*wzUser*/,
-        __in_z_opt LPCWSTR /*wzPassword*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SetVariableNumeric(
-        __in_z LPCWSTR /*wzVariable*/,
-        __in LONGLONG /*llValue*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SetVariableString(
-        __in_z LPCWSTR /*wzVariable*/,
-        __in_z_opt LPCWSTR /*wzValue*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP SetVariableVersion(
-        __in_z LPCWSTR /*wzVariable*/,
-        __in DWORD64 /*qwValue*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP CloseSplashScreen()
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP Detect(
-        __in_opt HWND /*hwndParent*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP Plan(
-        __in BOOTSTRAPPER_ACTION /*action*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP Elevate(
-        __in_opt HWND /*hwndParent*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP Apply(
-        __in_opt HWND /*hwndParent*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP Quit(
-        __in DWORD /*dwExitCode*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP LaunchApprovedExe(
-        __in_opt HWND /*hwndParent*/,
-        __in_z LPCWSTR /*wzApprovedExeForElevationId*/,
-        __in_z_opt LPCWSTR /*wzArguments*/,
-        __in DWORD /*dwWaitForInputIdleTimeout*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-public: // IMarshal
-    virtual STDMETHODIMP GetUnmarshalClass( 
-        __in REFIID /*riid*/,
-        __in_opt LPVOID /*pv*/,
-        __in DWORD /*dwDestContext*/,
-        __reserved LPVOID /*pvDestContext*/,
-        __in DWORD /*mshlflags*/,
-        __out LPCLSID /*pCid*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP GetMarshalSizeMax(
-        __in REFIID riid,
-        __in_opt LPVOID /*pv*/,
-        __in DWORD dwDestContext,
-        __reserved LPVOID /*pvDestContext*/,
-        __in DWORD /*mshlflags*/,
-        __out DWORD *pSize
-        )
-    {
-        HRESULT hr = S_OK;
-
-        // We only support marshaling the IBootstrapperEngine interface in-proc.
-        if (__uuidof(IBootstrapperEngine) != riid)
-        {
-            // Skip logging the following message since it appears way too often in the log.
-            // "Unexpected IID requested to be marshalled. BootstrapperEngineForApplication can only marshal the IBootstrapperEngine interface."
-            ExitFunction1(hr = E_NOINTERFACE);
-        }
-        else if (0 == (MSHCTX_INPROC & dwDestContext))
-        {
-            hr = E_FAIL;
-            ExitOnRootFailure(hr, "Cannot marshal IBootstrapperEngine interface out of proc.");
-        }
-
-        // E_FAIL is used because E_INVALIDARG is not a supported return value.
-        ExitOnNull(pSize, hr, E_FAIL, "Invalid size output parameter is NULL.");
-
-        // Specify enough size to marshal just the interface pointer across threads.
-        *pSize = sizeof(LPVOID);
-
-    LExit:
-        return hr;
-    }
-
-    virtual STDMETHODIMP MarshalInterface( 
-        __in IStream* pStm,
-        __in REFIID riid,
-        __in_opt LPVOID pv,
-        __in DWORD dwDestContext,
-        __reserved LPVOID /*pvDestContext*/,
-        __in DWORD /*mshlflags*/
-        )
-    {
-        HRESULT hr = S_OK;
-        IBootstrapperEngine *pThis = NULL;
-        ULONG ulWritten = 0;
-
-        // We only support marshaling the IBootstrapperEngine interface in-proc.
-        if (__uuidof(IBootstrapperEngine) != riid)
-        {
-            // Skip logging the following message since it appears way too often in the log.
-            // "Unexpected IID requested to be marshalled. BootstrapperEngineForApplication can only marshal the IBootstrapperEngine interface."
-            ExitFunction1(hr = E_NOINTERFACE);
-        }
-        else if (0 == (MSHCTX_INPROC & dwDestContext))
-        {
-            hr = E_FAIL;
-            ExitOnRootFailure(hr, "Cannot marshal IBootstrapperEngine interface out of proc.");
-        }
-
-        // "pv" may not be set, so we should us "this" otherwise.
-        if (pv)
-        {
-            pThis = reinterpret_cast<IBootstrapperEngine*>(pv);
-        }
-        else
-        {
-            pThis = static_cast<IBootstrapperEngine*>(this);
-        }
-
-        // E_INVALIDARG is not a supported return value.
-        ExitOnNull(pStm, hr, E_FAIL, "The marshaling stream parameter is NULL.");
-
-        // Marshal the interface pointer in-proc as is.
-        hr = pStm->Write(pThis, sizeof(pThis), &ulWritten);
-        if (STG_E_MEDIUMFULL == hr)
-        {
-            ExitOnFailure(hr, "Failed to write the stream because the stream is full.");
-        }
-        else if (FAILED(hr))
-        {
-            // All other STG error must be converted into E_FAIL based on IMarshal documentation.
-            hr = E_FAIL;
-            ExitOnFailure(hr, "Failed to write the IBootstrapperEngine interface pointer to the marshaling stream.");
-        }
-
-    LExit:
-        return hr;
-    }
-
-    virtual STDMETHODIMP UnmarshalInterface(
-        __in IStream* pStm,
-        __in REFIID riid,
-        __deref_out LPVOID* ppv
-        )
-    {
-        HRESULT hr = S_OK;
-        ULONG ulRead = 0;
-
-        // We only support marshaling the engine in-proc.
-        if (__uuidof(IBootstrapperEngine) != riid)
-        {
-            // Skip logging the following message since it appears way too often in the log.
-            // "Unexpected IID requested to be marshalled. BootstrapperEngineForApplication can only marshal the IBootstrapperEngine interface."
-            ExitFunction1(hr = E_NOINTERFACE);
-        }
-
-        // E_FAIL is used because E_INVALIDARG is not a supported return value.
-        ExitOnNull(pStm, hr, E_FAIL, "The marshaling stream parameter is NULL.");
-        ExitOnNull(ppv, hr, E_FAIL, "The interface output parameter is NULL.");
-
-        // Unmarshal the interface pointer in-proc as is.
-        hr = pStm->Read(*ppv, sizeof(LPVOID), &ulRead);
-        if (FAILED(hr))
-        {
-            // All STG errors must be converted into E_FAIL based on IMarshal documentation.
-            hr = E_FAIL;
-            ExitOnFailure(hr, "Failed to read the IBootstrapperEngine interface pointer from the marshaling stream.");
-        }
-
-    LExit:
-        return hr;
-    }
-
-    virtual STDMETHODIMP ReleaseMarshalData(
-        __in IStream* /*pStm*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-    virtual STDMETHODIMP DisconnectObject(
-        __in DWORD /*dwReserved*/
-        )
-    {
-        return E_NOTIMPL;
-    }
-
-public:
-    CEngineForApplication(
-        __in BURN_ENGINE_STATE* pEngineState,
-        __in DWORD dwThreadId
-        )
-    {
-        m_cReferences = 1;
-        m_pEngineState = pEngineState;
-        m_dwThreadId = dwThreadId;
-    }
-
-private:
-    long m_cReferences;
-    BURN_ENGINE_STATE* m_pEngineState;
-    DWORD m_dwThreadId;
-};
-
-
-extern "C" HRESULT EngineForApplicationCreate(
-    __in BURN_ENGINE_STATE* pEngineState,
-    __in DWORD dwThreadId,
-    __out IBootstrapperEngine** ppEngineForApplication
-    )
-{
-    HRESULT hr = S_OK;
-
-    CEngineForApplication* pEngine = new CEngineForApplication(pEngineState, dwThreadId);
-    ExitOnNull(pEngine, hr, E_OUTOFMEMORY, "Failed to allocate new BootstrapperEngineForApplication object.");
-
-    hr = pEngine->QueryInterface(IID_PPV_ARGS(ppEngineForApplication));
-    ExitOnFailure(hr, "Failed to QI for IBootstrapperEngine from BootstrapperEngineForApplication object.");
-
-LExit:
-    ReleaseObject(pEngine);
-    return hr;
-}
-
 HRESULT WINAPI EngineForApplicationProc(
     __in BOOTSTRAPPER_ENGINE_MESSAGE message,
     __in const LPVOID pvArgs,
