+++ src/burn/engine/precomp.h
@@ -58,8 +58,6 @@
 
 #include "BootstrapperEngine.h"
 #include "BootstrapperApplication.h"
-#include "IBootstrapperEngine.h"
-#include "IBootstrapperApplication.h"
 
 #include "platform.h"
 #include "variant.h"
